pyproject.toml
[tool.poetry]
name = "medium-backend"
version = "1.0.0"
description = "Medium App using Django & Graphql"
authors = ["Ramzi Tannous <r>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
graphene-django = "^2.15.0"
python-decouple = "^3.4"
dj-database-url = "^0.5.0"
psycopg2 = "^2.8.6"
Pillow = "^8.2.0"
django-graphql-jwt = "^0.3.1"
PyJWT = "^1.7.1"
graphene-django-optimizer = "^0.8.0"
django-filter = "^2.4.0"
django-debug-toolbar = "^3.2.1"
graphene-file-upload = "^1.3.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"
pytest = "^6.2.4"
pytest-django = "^4.2.0"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
